PLTCONN-3570: Rename classes to be more readable
