+++ lib/index.ts
@@ -6,6 +6,6 @@ export * from './connector'
 export * from './connector-customizer'
 export * from './error'
 export * from './handler'
-export * from './before-after-handlers'
+export * from './connector-customizer-handler'
 export * from './response'
 export * from './logger'
